+++ src/Migration/AbstractMigration.php
@@ -4,6 +4,7 @@ namespace Phoenix\Migration;
 
 use PDOStatement;
 use Phoenix\Database\Adapter\AdapterInterface;
+use Phoenix\Database\Element\ColumnSettings;
 use Phoenix\Database\Element\MigrationTable;
 use Phoenix\Database\Element\Structure;
 use Phoenix\Database\Element\Table;
@@ -187,20 +188,44 @@ abstract class AbstractMigration
         return $this;
     }
 
-    final protected function changeCollation(string $targetCollation): void
+    /**
+     * adds turn off checking foreign keys query to list of queries to execute
+     * @return void
+     */
+    final protected function checkForeignKeysOff(): void
     {
         $this->execute($this->adapter->buildDoNotCheckForeignKeysQuery());
+    }
+
+    /**
+     * adds turn on checking foreign keys query to list of queries to execute
+     * @return void
+     */
+    final protected function checkForeignKeysOn(): void
+    {
+        $this->execute($this->adapter->buildCheckForeignKeysQuery());
+    }
+
+    /**
+     * changes collation on all existing tables and columns
+     *
+     * @param string $targetCollation
+     * @return void
+     */
+    final protected function changeCollation(string $targetCollation): void
+    {
+        $this->checkForeignKeysOff();
         [$targetCharset,] = explode('_', $targetCollation, 2);
 
         foreach ($this->adapter->getStructure()->getTables() as $table) {
             $migrationTable = $this->table($table->getName(), $table->getPrimary(), $targetCharset, $targetCollation);
             foreach ($table->getColumns() as $column) {
                 $settings = $actualSettings = $column->getSettings()->getSettings();
-                if ($actualSettings['charset'] && $actualSettings['charset'] !== $targetCharset) {
-                    $settings['charset'] = $targetCharset;
+                if (isset($actualSettings[ColumnSettings::SETTING_CHARSET]) && $actualSettings[ColumnSettings::SETTING_CHARSET] !== $targetCharset) {
+                    $settings[ColumnSettings::SETTING_CHARSET] = $targetCharset;
                 }
-                if ($actualSettings['collation'] && $actualSettings['collation'] !== $targetCollation) {
-                    $settings['collation'] = $targetCollation;
+                if (isset($actualSettings[ColumnSettings::SETTING_COLLATION]) && $actualSettings[ColumnSettings::SETTING_COLLATION] !== $targetCollation) {
+                    $settings[ColumnSettings::SETTING_COLLATION] = $targetCollation;
                 }
                 if ($settings !== $actualSettings) {
                     $migrationTable->changeColumn($column->getName(), $column->getName(), $column->getType(), $settings);
@@ -208,12 +233,13 @@ abstract class AbstractMigration
             }
             $migrationTable->save();
         }
-        $this->execute($this->adapter->buildCheckForeignKeysQuery());
+        $this->checkForeignKeysOn();
     }
 
     /**
      * @param array $queries
      * @param bool $dry
+     * @return array
      *
      * @throws DatabaseQueryExecuteException
      */